+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Toggletip`: Improve styles.
 - Feature: `CheckboxButton`: Add component.
 - Feature: `CheckboxButtonSet`: Add component.
 - Feature: `DropdownSelect`: Change list style to cut last item if scrollable.